+++ R/print.cSEMAssess.R
@@ -306,31 +306,45 @@ print.cSEMAssess <- function(x, ...) {
     cat2("\n\n", rule2("Discriminant validity assessment"))
     # HTMT
     if(any(names(x) == "HTMT") && !is.null(x$HTMT)) {
-      cat2("\n\n\tHeterotrait-monotrait ratio of correlations matrix (HTMT matrix)\n\n")
+      header_htmt <- "Heterotrait-monotrait ratio of correlations matrix (HTMT matrix)"
+      cat2("\n\n\t", header_htmt,
+           "\n\t", makeLine(type = 1, width = nchar(header_htmt)), "\n",
+           "\n\tValues in the lower triangular part are the ", if(x$HTMT$absolute){"absolute "}, "HTMT values\n\n")
 
-      if(x$Information$.inference) {
-        cat2("\tValues in the upper triangular part are the ", 
-             paste0(100*(1 - x$Information$.alpha), "%-quantiles of the\n", 
-            "\tbootstrap confidence intervals (using .ci = '", x$Information$.ci, "')\n",
-            "\tbased on ", x$HTMT$nr_admissibles ," valid bootstrap runs.\n\n")) 
+      if(x$HTMT$inference != "none") {
+        if(x$HTMT$inference == "asymptotic") {
+          cat2("\tValues in the upper triangular part are the ",
+               paste0(100*(1 - x$Information$.alpha), "% limits of the\n",
+                      "\tasymptotic confidence intervals based on the delta-method.\n\n"))
+        } else if(x$HTMT$inference == "bootstrap") {
+          cat2("\tValues in the upper triangular part are the ",
+               paste0(100*(1 - x$Information$.alpha), "%-quantiles of the\n",
+              "\tbootstrap confidence intervals (using .ci = '", x$HTMT$ci, "')\n",
+              "\tbased on ", x$HTMT$nr_admissibles ," valid bootstrap runs.\n\n"))
+        }
       }
       print(x$HTMT$htmts)
     }
 # HTMT2
-    if(any(names(x) == "HTMT2") && !is.null(x$HTMT)) {
-      cat2("\n\n\tAdvanced heterotrait-monotrait ratio of correlations matrix (HTMT2 matrix)\n\n")
+    if(any(names(x) == "HTMT2") && !is.null(x$HTMT2)) {
+      header_htmt2 <- "Advanced heterotrait-monotrait ratio of correlations matrix (HTMT2 matrix)"
+      cat2("\n\n\t", header_htmt2,
+           "\n\t", makeLine(type = 1, width = nchar(header_htmt2)), "\n",
+           "\n\tValues in the lower triangular part are the ", if(x$HTMT2$absolute){"absolute "}, "HTMT2 values\n\n")
       
-      if(x$Information$.inference) {
+      if(x$HTMT2$inference == "bootstrap") {
         cat2("\tValues in the upper triangular part are the ", 
              paste0(100*(1 - x$Information$.alpha), "%-quantiles of the\n", 
-                    "\tbootstrap confidence intervals (using .ci = '", x$Information$.ci, "')\n",
+                    "\tbootstrap confidence intervals (using .ci = '", x$HTMT2$ci, "')\n",
                     "\tbased on ", x$HTMT2$nr_admissibles ," valid bootstrap runs.\n\n")) 
-      }
+      } 
       print(x$HTMT2$htmts)
     }    
 
     if(any(names(x) == "Fornell-Larcker")) {
-      cat2("\n\n\tFornell-Larcker matrix\n\n")
+      fornell_text <- "Fornell-Larcker matrix"
+      cat2("\n\n\t",fornell_text, 
+           "\n\t", makeLine(type = 1, width = nchar(fornell_text)), "\n")
       print(x$`Fornell-Larcker`)
     }
   }