+++ tests/testthat/test-assess.R
@@ -17,65 +17,67 @@ a <- assess(res, .only_common_factors = FALSE)
 
 test_that("Assess works for all choices of .quality_criterion", {
   expect_identical(class(a), "cSEMAssess")
-  expect_equivalent(a$AVE, c(0.4802837, 0.6318118, 0.5557484), tolerance = 1e-07)
-  expect_equivalent(a$AIC, c(-209.8370237, -318.7730588), tolerance = 1e-07)
-  expect_equivalent(a$AICc, c(292.2113634, 183.3077493), tolerance = 1e-07)
-  expect_equivalent(a$AICu, c(-207.8330130, -315.7640227), tolerance = 1e-07)
-  expect_equivalent(a$BIC, c(-201.4078075, -306.1292345), tolerance = 1e-07)
-  expect_equivalent(a$FPE, c(0.6572610489, 0.5285880026), tolerance = 1e-07)
-  expect_equivalent(a$GM, c(511.4292162, 517.6438243), tolerance = 1e-07)
-  expect_equivalent(a$HQ, c(-206.5294131, -313.8116428), tolerance = 1e-07)
-  expect_equivalent(a$HQc, c(-206.4704807, -313.7009215), tolerance = 1e-07)
-  expect_equivalent(a$Mallows_Cp, c(3, 5), tolerance = 1e-07)
-  expect_equivalent(a$RhoC, c(0.7339026, 0.8329352, 0.7883384), tolerance = 1e-07)
-  expect_equivalent(a$RhoC_mm, c(0.7341254, 0.9446712, 0.7876519), tolerance = 1e-07)
-  expect_equivalent(a$RhoC_weighted, c(0.7388285, 1.0000000, 0.7958698 ), tolerance = 1e-07)
-  expect_equivalent(a$RhoC_weighted_mm, c(0.7388285, 0.8845458, 0.7958698), tolerance = 1e-07)
-  expect_equivalent(a$DG, 0.005499167 , tolerance = 1e-07)
-  expect_equivalent(a$DL, 0.01041484, tolerance = 1e-07)
-  expect_equivalent(a$DML, 0.02928953, tolerance = 1e-07)
-  expect_equivalent(a$Df, 22, tolerance = 1e-07)
-  expect_equivalent(c(a$F2), c(0.53061872, 0.34344601, 0,
+  expect_equal(unname(a$AVE), c(0.4802837, 0.6318118, 0.5557484), tolerance = 1e-07)
+  expect_equal(unname(a$AIC), c(-209.8370237, -318.7730588), tolerance = 1e-07)
+  expect_equal(unname(a$AICc), c(292.2113634, 183.3077493), tolerance = 1e-07)
+  expect_equal(unname(a$AICu), c(-207.8330130, -315.7640227), tolerance = 1e-07)
+  expect_equal(unname(a$BIC), c(-201.4078075, -306.1292345), tolerance = 1e-07)
+  expect_equal(unname(a$FPE), c(0.6572610489, 0.5285880026), tolerance = 1e-07)
+  expect_equal(unname(a$GM), c(511.4292162, 517.6438243), tolerance = 1e-07)
+  expect_equal(unname(a$HQ), c(-206.5294131, -313.8116428), tolerance = 1e-07)
+  expect_equal(unname(a$HQc), c(-206.4704807, -313.7009215), tolerance = 1e-07)
+  expect_equal(unname(a$Mallows_Cp), c(3, 5), tolerance = 1e-07)
+  expect_equal(unname(a$RhoC), c(0.7339026, 0.8329352, 0.7883384), tolerance = 1e-07)
+  expect_equal(unname(a$RhoC_mm), c(0.7341254, 0.9446712, 0.7876519), tolerance = 1e-07)
+  expect_equal(unname(a$RhoC_weighted), c(0.7388285, 1.0000000, 0.7958698 ), tolerance = 1e-07)
+  expect_equal(unname(a$RhoC_weighted_mm), c(0.7388285, 0.8845458, 0.7958698), tolerance = 1e-07)
+  expect_equal(unname(a$DG), 0.005499167 , tolerance = 1e-07)
+  expect_equal(unname(a$DL), 0.010414836, tolerance = 1e-07)
+  expect_equal(unname(a$DML), 0.02928953, tolerance = 1e-07)
+  expect_equal(unname(a$Df), 22, tolerance = 1e-07)
+  expect_equal(unname(c(a$F2)), c(0.53061872, 0.34344601, 0,
                                0.06958792, 0.00000000, 0), 
                     tolerance = 1e-07)
-  expect_equivalent(a$Chi_square, 14.61547, tolerance = 1e-05)
-  expect_equivalent(a$Chi_square_df, 0.6643398, tolerance = 1e-07)
-  expect_equivalent(a$CFI, 1, tolerance = 1e-07)
-  expect_equivalent(a$GFI, 0.9927503, tolerance = 1e-07)
-  expect_equivalent(a$CN, 1159.24462058346, tolerance = 1e-07)
-  expect_equivalent(a$IFI, 1.005165, tolerance = 1e-06)
-  expect_equivalent(a$NFI, 0.9899318 , tolerance = 1e-07)
-  expect_equivalent(a$NNFI, 1, tolerance = 1e-06)
-  expect_equivalent(a$RMSEA, 0, tolerance = 1e-07)
-  expect_equivalent(a$RMS_theta, 0.08470806, tolerance = 1e-07)
-  expect_equivalent(a$SRMR, 0.01521318, tolerance = 1e-07)
-  expect_equivalent(c(a$`Fornell-Larcker`), c(0.4802837, 0.3466694, 0.4402532, 
+  expect_equal(unname(a$Chi_square), 14.61547, tolerance = 1e-05)
+  expect_equal(unname(a$Chi_square_df), 0.6643398, tolerance = 1e-07)
+  expect_equal(unname(a$CFI), 1, tolerance = 1e-07)
+  expect_equal(unname(a$GFI), 0.9927503, tolerance = 1e-07)
+  expect_equal(unname(a$CN), 1159.24462058346, tolerance = 1e-07)
+  expect_equal(unname(a$IFI), 1.005165, tolerance = 1e-06)
+  expect_equal(unname(a$NFI), 0.9899318 , tolerance = 1e-07)
+  expect_equal(unname(a$NNFI), 1, tolerance = 1e-06)
+  expect_equal(unname(a$RMSEA), 0, tolerance = 1e-07)
+  expect_equal(unname(a$RMS_theta), 0.08470806, tolerance = 1e-07)
+  expect_equal(unname(a$SRMR), 0.01521318, tolerance = 1e-06)
+  expect_equal(unname(c(a$`Fornell-Larcker`)), c(0.4802837, 0.3466694, 0.4402532, 
                                               0.3466694, 0.6318118, 0.2969352,
                                               0.4402532, 0.2969352, 0.5557484), 
                     tolerance = 1e-07)
-  expect_equivalent(a$GoF, 0.4784006 , tolerance = 1e-07)
-  expect_equivalent(c(a$HTMT$htmts), c(1, 0.6782752, 0.6668841,
+  expect_equal(unname(a$GoF), 0.4784006 , tolerance = 1e-07)
+  expect_equal(unname(c(a$HTMT$htmts)), c(1, 0.6782752, 0.6668841,
                                  0, 1.0000000, 0.6124418,
                                  0, 0.0000000, 1.0000000), tolerance = 1e-07)
-  expect_equivalent(c(a$HTMT2$htmts), c(1.0000000, 0.6724003, 0.6652760, 
+  expect_equal(unname(c(a$HTMT2$htmts)), c(1.0000000, 0.6724003, 0.6652760, 
                                          0.0000000, 1.0000000, 0.5958725,
                                          0.0000000, 0.0000000, 1.0000000), tolerance = 1e-07)
-  expect_equivalent(a$R2, c(0.3466694, 0.4766706), tolerance = 1e-07)
-  expect_equivalent(a$R2_adj, c(0.3453575, 0.4745646), tolerance = 1e-07)
-  expect_equivalent(a$RhoT, c(0.7317595, 0.7280738, 0.7859542), tolerance = 1e-07)
-  expect_equivalent(a$RhoT_weighted, c(0.7288400, 0.6637134, 0.7821770) , tolerance = 1e-07)
-  expect_equivalent(c(a$VIF), c(1.530619, 1.530619), tolerance = 1e-06)
-  expect_equivalent(c(a$VIF_modeB), c(1.271908, 1.619737, 1.620424), tolerance = 1e-06)
+  expect_equal(unname(a$R2), c(0.3466694, 0.4766706), tolerance = 1e-07)
+  expect_equal(unname(a$R2_adj), c(0.3453575, 0.4745646), tolerance = 1e-07)
+  expect_equal(unname(a$RhoT), c(0.7317595, 0.7280738, 0.7859542), tolerance = 1e-07)
+  expect_equal(unname(a$RhoT_weighted), c(0.7288400, 0.6637134, 0.7821770) , tolerance = 1e-07)
+  expect_equal(unname(c(a$VIF)), c(1.530619, 1.530619), tolerance = 1e-06)
+  expect_equal(unname(c(a$VIF_modeB)), c(1.271908, 1.619737, 1.620424), tolerance = 1e-06)
 })
 
 ## Assess using additional arguments
+## .inference is set to "bootstrap" for tests regarding the asymptotic pipeline, 
+## see test-helper_assess.R
 assess(res,   
             .absolute            = TRUE,
             .alpha               = 0.05,
             .ci                  = "CI_percentile",
             .closed_form_ci      = FALSE,
             .handle_inadmissibles= "drop",
-            .inference           = TRUE,
+            .inference           = "bootstrap",
             .null_model          = FALSE,
             .R                   = 199,
             .saturated           = FALSE,