+++ NEWS.md
@@ -2,6 +2,8 @@
 
 #  cSEM 0.6.2
 
+- Extend calculateHTMT() to allow for asymptotic inference for the HTMT. Thanks to Jason Berger for his contribution.
+
 - Fix bug in calculating the moments used to estimate non-linear models. 
 
 - Fix bug in testMICOM(). Thanks to manzi0 for this contribution!